refactor(project): Express 프로젝트 구조 설정 - 빈 파일로 생성
