+++ package.json
@@ -34,6 +34,6 @@
     "typescript": "^5.8.2"
   },
   "prisma": {
-    "schema": "src/db/prisma/schema.prisma"
+    "schema": "src/models/prisma/schema.prisma"
   }
 }
