--- src/db/prisma/client.prisma.js
@@ -1,5 +1,0 @@
-const { PrismaClient } = require("@prisma/client");
-
-const prisma = new PrismaClient();
-
-module.exports = prisma;
