+++ src/models/prisma/prismaClient.js
@@ -1,0 +1,5 @@
+const { PrismaClient } = require("@prisma/client");
+
+const prismaClient = new PrismaClient();
+
+module.exports = prismaClient;
